The updater delegate can now add a custom update validation error
This allows you to reject an update with an error message depending on conditions that you determine.

This can be used for example, to reject updates that would update to a new major release of the application: You can reject major version updates and then let Sparkle show an error message that a new major update is available that has to be manually installed by the user.